now using OpenSSL for certificate parsing
`s.getpeercert()` returns an empty dict when cert validation fails, but
I can apparently always get a binary representation.

I feed this binary representation to OpenSSL and iterate over the
extensions to find the subjectAltName. I failed to find an easy way to
deserialize the ASN.1 sequence, so I use the dirty hack of butchering
the string representation to extract what I want to know.